requirements.txt
advent-cli>=0.1.1
advent-of-code-data>=2.0.1
aiohappyeyeballs>=2.4.3
aiohttp>=3.10.10
aiosignal>=1.3.1
aocd>=0.1
aocd-example-parser>=2023.12.17
attrs>=24.2.0
audioop-lts>=0.2.1
beautifulsoup4>=4.12.3
black>=24.10.0
certifi>=2024.8.30
charset-normalizer>=3.4.0
click>=8.1.7
colorama>=0.4.6
discord.py>=2.4.0
frozenlist>=1.4.1
idna>=3.10
markdownify>=0.13.1
multidict>=6.1.0
mypy-extensions>=1.0.0
packaging>=24.1
pathspec>=0.12.1
Pebble>=5.0.7
pipdeptree>=2.23.4
platformdirs>=4.3.6
propcache>=0.2.0
pytz>=2024.2
requests>=2.32.3
six>=1.16.0
soupsieve>=2.6
tabulate>=0.9.0
termcolor>=2.5.0
toml>=0.10.2
tzdata>=2024.2
urllib3>=2.2.3
yarl>=1.15.5
pytest>=7